more work on responsiveness of sidebar
