--- src/components/PhotoGallery/index.js
@@ -1,2 +1,0 @@
-import PhotoGallery from './PhotoGallery'
-export default PhotoGallery
